fix argument parsing for real
some of the issues were caused by the arguments going out of scope by
the time the callback is ran, causing general weirdness. this is my
fault, and is the main fix here

newer versions of CLI11 however also do weird stuff (like splitting
single arguments like '[1,2]' up into two arguments, with no way to
disable it. why?), so that is reverted in a previous commit too

also fixed the naming of some copy-pasta'd sub-command variables